Rename to flight-server to distinguish from the client parts
