+++ packages/react-dom/npm/unstable-flight.browser.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-dom-unstable-flight.browser.production.min.js');
+  module.exports = require('./cjs/react-dom-unstable-flight-server.browser.production.min.js');
 } else {
-  module.exports = require('./cjs/react-dom-unstable-flight.browser.development.js');
+  module.exports = require('./cjs/react-dom-unstable-flight-server.browser.development.js');
 }
