+++ packages/react-dom/npm/unstable-flight.js
@@ -1,3 +1,3 @@
 'use strict';
 
-module.exports = require('./unstable-flight.node');
+module.exports = require('./unstable-flight-server.node');
