+++ packages/react-dom/npm/unstable-flight.node.js
@@ -1,7 +1,7 @@
 'use strict';
 
 if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./cjs/react-dom-unstable-flight.node.production.min.js');
+  module.exports = require('./cjs/react-dom-unstable-flight-server.node.production.min.js');
 } else {
-  module.exports = require('./cjs/react-dom-unstable-flight.node.development.js');
+  module.exports = require('./cjs/react-dom-unstable-flight-server.node.development.js');
 }
