+++ packages/react-dom/unstable-flight-server.js
@@ -9,4 +9,4 @@
 
 'use strict';
 
-module.exports = require('./unstable-flight.node');
+module.exports = require('./unstable-flight-server.node');
