+++ packages/react-noop-renderer/flight-server.js
@@ -9,8 +9,8 @@
 
 'use strict';
 
-const ReactFlightDOMBrowser = require('./src/server/flight/ReactFlightDOMBrowser');
+const ReactNoopFlightServer = require('./src/ReactNoopFlightServer');
 
 // TODO: decide on the top-level export form.
-// This is hacky but makes it work with both Rollup and Jest
-module.exports = ReactFlightDOMBrowser.default || ReactFlightDOMBrowser;
+// This is hacky but makes it work with both Rollup and Jest.
+module.exports = ReactNoopFlightServer.default || ReactNoopFlightServer;
